+++ CHANGELOG.rst
@@ -21,7 +21,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 - `UpdateTZData` command added
 
   - Downloads the latest TZ data file from the TZ_DATA_URL env var and uploads it to SPFRx
+
 - Update receiver band to Band1 for SPFRx if requested band is B5b
+- Add param for delay to FannedOutCommand completion. Used for SPFRx ConfigureBand command.
 
 Version 10.0.0
 **************